Add profile routing and navigation
A new "Profile" route has been added to the routing system and an associated "My Profile" navigation item has been created in the side bar. The new "ProfileView" file has been created to support this and a navigational button has been activated on the "ResultView" page for redirecting to the "Profile" page.