Update SideBar component and use in ProfileView
The SideBar component was updated to have more dynamic data, such as a configurable header name and icon. It was also improved to cater for optional progress indication. The ProfileView was updated to incorporate the use of the SideBar component. Updates also include minor layout adjustments and addition of a new SVG asset.