Add QuizCertificate Vue.js component and SCSS file
This commit introduces a new Vue.js component, QuizCertificate, specifically designed to render quiz completion certificates. A matching SCSS file is also added to handle the certificate's styling. This component is now imported and used in the ProfileView component. A new SVG file for a stamp graphic displayed on the certificate is also included.